Created A simplistic model (#6)
The complete model that holds Infinion's physics is causing too many problems at the moment. In an effort to not hold development back, I've created a simplistic model of an airplane. It's not perfect, so the PID gains derived will not be valid for Spike/Infinion, but it will serve the purpose of allowing us to debug the higher level functionality of the autopilot. Things like cruising, take off, landing, and path management.

The big difference is that this simplistic model does not deal with Datcom at all. Instead, simple physics are done.

Also, this model has exposed flaws in the real model. Flaws such as misunderstandings of coordinate frames. They will be fixed soon.